move tagging logic to abstract
