+++ src/main/java/uk/gov/hmcts/reform/preapi/media/storage/AzureIngestStorageService.java
@@ -1,15 +1,9 @@
 package uk.gov.hmcts.reform.preapi.media.storage;
 
-import com.azure.storage.blob.BlobClient;
 import com.azure.storage.blob.BlobServiceClient;
-import com.azure.storage.blob.models.BlobItem;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.HashMap;
-
-@Slf4j
 @Service
 public class AzureIngestStorageService extends AzureStorageService {
     @Autowired
@@ -27,22 +21,5 @@ public class AzureIngestStorageService extends AzureStorageService {
 
     public void markContainerAsSafeToDelete(String containerName) {
         tagAllBlobsInContainer(containerName, "status", "safe_to_delete");
-    }
-
-    protected void tagAllBlobsInContainer(String containerName, String tagKey, String tagValue) {
-        log.info("Setting all blob's tags in container '{}': '{}' to '{}'", containerName, tagKey, tagValue);
-        var containerClient = client.getBlobContainerClient(containerName);
-        containerClient.listBlobs()
-            .stream()
-            .map(BlobItem::getName)
-            .map(containerClient::getBlobClient)
-            .forEach(blob -> setBlobTag(blob, tagKey, tagValue));
-    }
-
-    protected void setBlobTag(BlobClient blob, String tagKey, String tagValue) {
-        var currentTags = blob.getTags();
-        var newTags = new HashMap<>(currentTags);
-        newTags.put(tagKey, tagValue);
-        blob.setTags(newTags);
     }
 }
